requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
azure-core==1.25.0
azure-storage-blob==12.13.1
bcrypt @ file:///C:/Windows/Temp/abs_36kl66t_aw/croots/recipe/bcrypt_1659554334050/work
certifi==2022.6.15
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer==2.1.0
cryptography @ file:///C:/ci/cryptography_1652101772384/work
defusedxml==0.7.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-filter==22.1
django-sslserver==0.22
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
isodate==0.6.1
msrest==0.7.1
numpy @ file:///D:/bld/numpy_1660672939211/work
oauthlib==3.2.0
opencv-python==4.6.0
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT==1.7.1
PyMySQL @ file:///C:/ci/pymysql_1610482945113/work
python3-openid==3.2.0
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
requests==2.28.1
requests-oauthlib==1.3.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==1.26.11
wincertstore==0.2
